+++ docs/pages/components/notification/Notification.vue
@@ -10,6 +10,10 @@
 
         <Example :component="ExIcons" :code="ExIconsCode" title="Icons" vertical/>
 
+        <Example :component="ExAutoClose" :code="ExAutoCloseCode" title="Auto Close" vertical>
+            <p>Notification will be automatically closed after <code>duration</code>.</p>
+        </Example>
+
         <ApiView :data="api"/>
     </div>
 </template>
@@ -26,6 +30,9 @@
     import ExIcons from './examples/ExIcons'
     import ExIconsCode from '!!raw-loader!./examples/ExIcons'
 
+    import ExAutoClose from './examples/ExAutoClose'
+    import ExAutoCloseCode from '!!raw-loader!./examples/ExAutoClose'
+
     export default {
         data() {
             return {
@@ -33,9 +40,11 @@
                 ExSimple,
                 ExTypes,
                 ExIcons,
+                ExAutoClose,
                 ExSimpleCode,
                 ExTypesCode,
-                ExIconsCode
+                ExIconsCode,
+                ExAutoCloseCode
             }
         }
     }