+++ src/components/message/Message.vue
@@ -37,45 +37,10 @@
     export default {
         name: 'BMessage',
         mixins: [MessageMixin],
-        props: {
-            autoClose: {
-                type: Boolean,
-                default: false
-            },
-            duration: {
-                type: Number,
-                default: 5000
-            }
-        },
         data() {
             return {
                 newIconSize: this.iconSize || this.size || 'is-large'
             }
-        },
-        watch: {
-            isActive(value) {
-                if (value) {
-                    this.setAutoClose()
-                } else {
-                    if (this.timer) {
-                        clearTimeout(this.timer)
-                    }
-                }
-            }
-        },
-        methods: {
-            setAutoClose() {
-                if (this.autoClose) {
-                    this.timer = setTimeout(() => {
-                        if (this.isActive) {
-                            this.close()
-                        }
-                    }, this.duration)
-                }
-            }
-        },
-        mounted() {
-            this.setAutoClose()
         }
     }
 </script>